refactor reused normalizeValue
