+++ lib/options/format.js
@@ -1,6 +1,7 @@
 var systemLineBreak = require('os').EOL;
 
 var override = require('../utils/override');
+var normalizeValue = require('../utils/normalize-value');
 
 var Breaks = {
   AfterAtRule: 'afterAtRule',
@@ -48,11 +49,6 @@ var OPTION_SEPARATOR = ';';
 var OPTION_NAME_VALUE_SEPARATOR = ':';
 var HASH_VALUES_OPTION_SEPARATOR = ',';
 var HASH_VALUES_NAME_VALUE_SEPARATOR = '=';
-
-var FALSE_KEYWORD_1 = 'false';
-var FALSE_KEYWORD_2 = 'off';
-var TRUE_KEYWORD_1 = 'true';
-var TRUE_KEYWORD_2 = 'on';
 
 function breaks(value) {
   var breakOptions = {};
@@ -169,19 +165,6 @@ function hashValuesToHash(string) {
 
       return accumulator;
     }, {});
-}
-
-function normalizeValue(value) {
-  switch (value) {
-  case FALSE_KEYWORD_1:
-  case FALSE_KEYWORD_2:
-    return false;
-  case TRUE_KEYWORD_1:
-  case TRUE_KEYWORD_2:
-    return true;
-  default:
-    return value;
-  }
 }
 
 function mapBreakWith(value) {
