+++ lib/options/optimization-level.js
@@ -1,6 +1,7 @@
 var roundingPrecisionFrom = require('./rounding-precision').roundingPrecisionFrom;
 
 var override = require('../utils/override');
+var normalizeValue = require('../utils/normalize-value');
 
 var OptimizationLevel = {
   Zero: '0',
@@ -53,10 +54,6 @@ DEFAULTS[OptimizationLevel.Two] = {
 
 var ALL_KEYWORD_1 = '*';
 var ALL_KEYWORD_2 = 'all';
-var FALSE_KEYWORD_1 = 'false';
-var FALSE_KEYWORD_2 = 'off';
-var TRUE_KEYWORD_1 = 'true';
-var TRUE_KEYWORD_2 = 'on';
 
 var LIST_VALUE_SEPARATOR = ',';
 var OPTION_SEPARATOR = ';';
@@ -154,19 +151,6 @@ function defaults(level, value) {
   }
 
   return options;
-}
-
-function normalizeValue(value) {
-  switch (value) {
-  case FALSE_KEYWORD_1:
-  case FALSE_KEYWORD_2:
-    return false;
-  case TRUE_KEYWORD_1:
-  case TRUE_KEYWORD_2:
-    return true;
-  default:
-    return value;
-  }
 }
 
 function convertValuesToHashes(source) {
